requirements.txt
requests
PyYAML
colorama
